+++ pure_extensions/lib/built_extensions.dart
@@ -6,4 +6,5 @@ export 'src/built/geo_point_serializer.dart';
 export 'src/built/iterable.dart';
 export 'src/built/locale_serializer.dart';
 export 'src/built/rational_serializer.dart';
+export 'src/built/raw_serializer.dart';
 export 'src/built/serializers.dart';