+++ pure_extensions/lib/src/dart/primitives/geo_point.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 
+import 'package:pure_extensions/src/dart/collections/iterable.dart';
 import 'package:pure_extensions/src/dart/primitives/numbers.dart';
 
 /// Defines a geographical point
@@ -11,6 +12,7 @@ class GeoPoint {
 
   static const double earthRadius = 6371000.0;
 
+  /// Returns a distance
   double distance(GeoPoint other, {double radius = GeoPoint.earthRadius}) {
     final sDLat = sin((other.latitude.toRad() - latitude.toRad()) / 2);
     final sDLng = sin((other.longitude.toRad() - longitude.toRad()) / 2);
@@ -73,7 +75,7 @@ class GeoPoint {
 
   @override
   String toString() {
-    return 'GeoPoint{latitude: $latitude, longitude: $longitude}';
+    return 'GeoPoint($latitude, $longitude)';
   }
 }
 
@@ -83,6 +85,11 @@ class GeoBounds {
   final GeoPoint southwest;
 
   const GeoBounds({required this.northeast, required this.southwest});
+
+  /// Returns a distance
+  double distance({double radius = GeoPoint.earthRadius}) {
+    return northeast.distance(southwest, radius: radius);
+  }
 
   /// Returns whether this rectangle contains the given [LatLng].
   bool contains(GeoPoint point) {
@@ -136,7 +143,48 @@ class GeoBounds {
   int get hashCode => northeast.hashCode ^ southwest.hashCode;
 
   @override
-  String toString() {
-    return 'GeoBounds{northeast: $northeast, southwest: $southwest}';
+  String toString() => 'GeoBounds(northeast: $northeast, southwest: $southwest}';
+}
+
+extension IterableGeoPointDartExtension on Iterable<GeoPoint> {
+  /// Calculate the distance between geographic points.
+  double geoDistance<T>() => foldWithNext(0.0, (count, p, c) => count + p.distance(c));
+
+  /// Calculate a center.
+  GeoPoint geoCenter() {
+    final eb = this.externalGeoBounds();
+    final p = eb.northeast + eb.southwest;
+    return GeoPoint(p.latitude / 2, p.longitude / 2);
+  }
+
+  /// Calculate the northeast corner.
+  GeoPoint geoNortheast() {
+    return reduce((p, c) {
+      return GeoPoint(max(p.latitude, c.latitude), max(p.longitude, c.longitude));
+    });
+  }
+
+  /// Calculate the southwest corner.
+  GeoPoint geoSouthwest() {
+    return reduce((p, c) {
+      return GeoPoint(min(p.latitude, c.latitude), min(p.longitude, c.longitude));
+    });
+  }
+
+  /// Calculate the internal corners.
+  GeoBounds internalGeoBounds({GeoPoint? center}) {
+    center ??= geoCenter();
+    return GeoBounds(
+      northeast: where((p) => p >= center!).geoSouthwest(),
+      southwest: where((p) => p <= center!).geoNortheast(),
+    );
+  }
+
+  /// Calculate the external corners.
+  GeoBounds externalGeoBounds() {
+    return GeoBounds(
+      northeast: geoNortheast(),
+      southwest: geoSouthwest(),
+    );
   }
 }